requirements.txt
# prod


python-dotenv==0.15.0
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.0.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0


# dev and test
black==19.3b0
flake8==3.8.3
isort==5.4.2